// src/api/types/PositionDto.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

export type PositionDto = {
    referenceId?: string;
    name?: string;
    code?: string;
    description?: string | null;
    /**
     * @description
     * Format: `date-time`
     * @type string | undefined
    */
    createdOnUtc?: string;
    createdBy?: string | null;
    /**
     * @description
     * Format: `date-time`
     * @type string | undefined
    */
    lastModifiedOnUtc?: string | null;
    lastModifiedBy?: string | null;
    isDeleted?: boolean;
    /**
     * @description
     * Format: `date-time`
     * @type string | undefined
    */
    deletedOnUtc?: string | null;
    deletedBy?: string | null;
};